fix votekick + kick exploit [original aloha commit: 5586684, BR]
prior to the commit, votekicks were cancelled when the "victim" was kicked
this means that if you are about to be votekicked, and intentionally kick yourself,
you could end the votekick without being banned for 30 minutes.
an easy way to kick yourself is to /login with the wrong password a few times.

this commit makes it so if you are kicked with a votekick against you,
the votekick will not be cancelled and will apply the "disconnect" ban.

this also fixes a long-standing bug in on_ban, and will now cancel the votekick
if the player is /banned. (previously it fell back to the kick handler)
this does not apply to /banip so be careful.
if you /banip someone who is connected with a votekick against them.
they will end up banned twice, and I seem to remember that when the shorter ban
duration expires, the longer ban is cleared also.