Update arena.py
The scoreboard in the arena game mode should represent the actual kill counts to avoid confusions. Therefore this overwrites Piqueservers capture_flag logic and sends an ExistingPlayer packet to all players so they can see the actual kill count.

For OpenSpades and BetterSpades it works so far as intended. Voxlap (vanilla) joins with seeing the real kill counts of the game but each intel capture still results in +10 kills on the client.